pyproject.toml
[project]
name = "scc-torres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.121.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.44",
]
